Handle logs with NUL chars better
It is not too rare that '\0' chars appear in /var/log/messages. I saw a
real-life case where kerneloops would show a popup with the same old oops
after every login. It turned out that there were NUL chars in the log which
prevented kerneloops from seeing its marker, so it always treated the old oops
in the log as new.
This patch fixes it by always going through the whole known length of the
buffer (not stopping on NUL chars) and using less string-oriented functions in
fill_lineinfo().